requirements.txt
fastapi
uvicorn
pydantic_settings
python-dotenv
requests
websockets
pymysql
sqlalchemy